+++ src/subsettools/subsetting.py
@@ -12,11 +12,13 @@ import os
 from datetime import datetime, timedelta
 import threading
 import re
+import warnings
 import numpy as np
 import hf_hydrodata
 from parflow.tools.io import write_pfb
 from ._common import (
     get_utc_time,
+    get_hf_gridded_data,
 )
 from ._error_checking import (
     _validate_grid,
@@ -52,14 +54,15 @@ def subset_static(
         - Subsurface units indicator file (pf_indicator)
         - Mannings roughness coefficients (mannings)
         - Depth to bedrock (pf_flowbarrier)
-        - Long term average preciptation minus evaporation (i.e. recharge) (pme)
+        - Long term average precipitation minus evaporation (i.e. recharge) (pme)
         - Steady state pressure head used to initialize transient simulations
           (ss_pressure_head)
 
     Note that some datasets might not contain all 7 static input variables. In
-    that case, the subset_static function is going to get and save the data for
-    the variables supported by the dataset and print out a message for those
-    that are not.
+    that case, the subset_static function is going to raise a ValueError for any
+    variables that do not exist in the dataset. The default variable list
+    contains the necessary static variables for the CONUS2 grid. For CONUS1-based
+    datasets, "mannings" and "pf_flowbarrier" should be removed from the list.
 
     Args:
         ij_bounds (tuple[int]): bounding box for subset. This should be given as
@@ -71,7 +74,7 @@ def subset_static(
         write_dir (str): directory where the subset files will be written
         var_list (tuple[str]): tuple of variables to subset from the dataset.
             By default all 7 variables above will be subset. The user can specify
-            as subset of these variables or list additional variables that are
+            a subset of these variables or list additional variables that are
             available in their dataset of choice.
 
     Returns:
@@ -82,13 +85,34 @@ def subset_static(
 
     .. code-block:: python
 
+        # Subsetting static variables for a CONUS1 workflow
+        # We need to remove "pf_flowbarrier" and "mannings" from the list
         filepaths = subset_static(
             ij_bounds=(375, 239, 487, 329),
             dataset="conus1_domain",
             write_dir="/path/to/your/chosen/directory",
-            var_list=("slope_x", "slope_y")
+            var_list=("slope_x", "slope_y", "pf_indicator", "pme",
+                      "ss_pressure_head")
+        )
+
+        # Subsetting static variables for a CONUS2 workflow
+        # Note that we can use the default var_list here
+        filepaths = subset_static(
+            ij_bounds=(3701, 1544, 3792, 1633),
+            dataset="conus2_domain",
+            write_dir="/path/to/your/chosen/directory",
         )
     """
+    warnings.warn(
+        "Note that for subsettools versions >= 2.0.0, this function will raise "
+        "a ValueError if a variable in var_list is not supported in the "
+        "dataset. (In older versions, it just printed an error message and "
+        "continued executing normally). You can check in the HydroData "
+        "documentation which variables are contained in each dataset "
+        "(https://hf-hydrodata.readthedocs.io/en/latest/available_data.html).",
+        DeprecationWarning,
+        stacklevel=2,
+    )
     _validate_grid_bounds(ij_bounds)
     if not isinstance(dataset, str):
         raise TypeError("dataset name must be a string.")
@@ -96,22 +120,19 @@ def subset_static(
     if not all(isinstance(var, str) for var in var_list):
         raise TypeError("All variable names should be strings.")
     file_paths = {}
+    options = {
+        "dataset": dataset,
+        "file_type": "pfb",
+        "temporal_resolution": "static",
+        "grid_bounds": ij_bounds,
+    }
     for var in var_list:
-        try:
-            subset_data = hf_hydrodata.get_gridded_data(
-                dataset=dataset,
-                variable=var,
-                file_type="pfb",
-                temporal_resolution="static",
-                grid_bounds=ij_bounds,
-            )
-        except ValueError as err:
-            print(f"Variable '{var}' not found in dataset '{dataset}':", err)
-        else:
-            file_path = os.path.join(write_dir, f"{var}.pfb")
-            write_pfb(file_path, subset_data, dist=False)
-            file_paths[var] = file_path
-            print(f"Wrote {var}.pfb in specified directory.")
+        options["variable"] = var
+        subset_data = get_hf_gridded_data(options)
+        file_path = os.path.join(write_dir, f"{var}.pfb")
+        write_pfb(file_path, subset_data, dist=False)
+        file_paths[var] = file_path
+        print(f"Wrote {var}.pfb in specified directory.")
 
     return file_paths
 
@@ -165,19 +186,15 @@ def subset_press_init(ij_bounds, dataset, date, write_dir, time_zone="UTC"):
     print(f"UTC Date: {new_date}")
     date_string = new_date.strftime("%Y.%m.%d:%H.%M.%S_UTC0")
 
-    try:
-        subset_data = hf_hydrodata.get_gridded_data(
-            dataset=dataset,
-            variable="pressure_head",
-            file_type="pfb",
-            temporal_resolution="hourly",
-            grid_bounds=ij_bounds,
-            start_time=new_date,
-        )
-    except ValueError as err:
-        print(f"No pressure file found in dataset '{dataset}':", err)
-        return None
-
+    options = {
+        "dataset": dataset,
+        "variable": "pressure_head",
+        "file_type": "pfb",
+        "temporal_resolution": "hourly",
+        "grid_bounds": ij_bounds,
+        "start_time": new_date,
+    }
+    subset_data = get_hf_gridded_data(options)
     file_path = os.path.join(write_dir, f"{dataset}_{date_string}_press.pfb")
     write_pfb(file_path, subset_data[0, :, :, :], dist=False)
     print(f"Wrote {file_path} in specified directory.")
@@ -262,10 +279,12 @@ def subset_forcing(
         raise TypeError("time_zone must be a string.")
     if not all(isinstance(var, str) for var in forcing_vars):
         raise TypeError("All variable names should be strings.")
+    forcing_vars = tuple(set(forcing_vars))
     outputs = {}
     start_date = get_utc_time(start, time_zone)
     end_date = get_utc_time(end, time_zone)
     exit_event = threading.Event()
+    lock = threading.Lock()
     threads = [
         threading.Thread(
             target=_subset_forcing_variable,
@@ -280,6 +299,7 @@ def subset_forcing(
                 time_zone,
                 outputs,
                 exit_event,
+                lock,
             ),
         )
         for variable in forcing_vars
@@ -294,13 +314,13 @@ def subset_forcing(
     except KeyboardInterrupt:
         print("Interrupted. Stopping threads...")
         exit_event.set()
-
         for thread in threads:
             thread.join()
-
         print("All threads stopped.")
-    finally:
-        return outputs
+
+    if exit_event.is_set():
+        raise ValueError("One or more threads were interrupted.")
+    return outputs
 
 
 def _subset_forcing_variable(
@@ -314,14 +334,23 @@ def _subset_forcing_variable(
     time_zone,
     outputs,
     exit_event,
+    lock,
 ):
     """Helper for subset_forcing that subsets data for one forcing variable."""
 
     day = 1
     date = start_date
     delta = timedelta(hours=_HOURS_PER_FORCING_FILE)
-    outputs[variable] = []
+    write_paths = []
     print(f"Reading {variable} pfb sequence")
+    options = {
+        "dataset": dataset,
+        "variable": variable,
+        "grid": grid,
+        "file_type": "pfb",
+        "temporal_resolution": "hourly",
+        "grid_bounds": ij_bounds,
+    }
 
     while date < end_date and not exit_event.is_set():
         start_time = date
@@ -331,64 +360,34 @@ def _subset_forcing_variable(
         # middle of a file and span multiple files
         try:
             if time_zone == "UTC":
-                subset_data = hf_hydrodata.get_gridded_data(
-                    dataset=dataset,
-                    variable=variable,
-                    grid=grid,
-                    file_type="pfb",
-                    temporal_resolution="hourly",
-                    start_time=start_time,
-                    end_time=end_time,
-                    grid_bounds=ij_bounds,
-                )
+                options["start_time"] = start_time
+                options["end_time"] = end_time
+                subset_data = get_hf_gridded_data(options)
             else:
                 next_day_midnight = datetime(
                     end_time.year, end_time.month, end_time.day
                 )
-                data1 = hf_hydrodata.get_gridded_data(
-                    dataset=dataset,
-                    variable=variable,
-                    grid=grid,
-                    file_type="pfb",
-                    temporal_resolution="hourly",
-                    start_time=start_time,
-                    end_time=next_day_midnight,
-                    grid_bounds=ij_bounds,
-                )
-                data2 = hf_hydrodata.get_gridded_data(
-                    dataset=dataset,
-                    variable=variable,
-                    grid=grid,
-                    file_type="pfb",
-                    temporal_resolution="hourly",
-                    start_time=next_day_midnight,
-                    end_time=end_time,
-                    grid_bounds=ij_bounds,
-                )
-                subset_data = np.concatenate((data1, data2), axis=0)
-        except Exception as exc:
-            raise ValueError(
-                f"Failed to get {variable} data from {start_time} to {end_time}."
-            ) from exc
-
-        paths = hf_hydrodata.get_paths(
-            dataset=dataset,
-            variable=variable,
-            grid=grid,
-            file_type="pfb",
-            temporal_resolution="hourly",
-            start_time=start_time,
-            end_time=end_time,
-        )
-
+                options["start_time"] = start_time
+                options["end_time"] = next_day_midnight
+                data_day1 = get_hf_gridded_data(options)
+                options["start_time"] = next_day_midnight
+                options["end_time"] = end_time
+                data_day2 = get_hf_gridded_data(options)
+                subset_data = np.concatenate((data_day1, data_day2), axis=0)
+        except Exception:
+            exit_event.set()
+            raise
+        paths = hf_hydrodata.get_paths(options)
         write_path = os.path.join(
             write_dir, _adjust_filename_hours(os.path.basename(paths[0]), day)
         )
-        outputs[variable].append(write_path)
+        write_paths.append(write_path)
         write_pfb(write_path, subset_data[:, :, :], dist=False)
         date = date + delta
         day = day + 1
     if not exit_event.is_set():
+        with lock:
+            outputs[variable] = write_paths
         print(f"Finished writing {variable} to folder")