+++ tests/test_conus1_upper_verde_spinup.py
@@ -28,7 +28,18 @@ def test_conus1_upper_verde_spinup(setup_dir_structure, remove_output_files):
 
     ij_bounds, mask = st.define_huc_domain(hucs=hucs, grid=grid)
     st.write_mask_solid(mask=mask, grid=grid, write_dir=static_write_dir)
-    st.subset_static(ij_bounds, dataset=var_ds, write_dir=static_write_dir)
+    st.subset_static(
+        ij_bounds,
+        dataset=var_ds,
+        write_dir=static_write_dir,
+        var_list=(
+            "slope_x",
+            "slope_y",
+            "pf_indicator",
+            "pme",
+            "ss_pressure_head",
+        ),
+    )
     init_press_filepath = st.subset_press_init(
         ij_bounds,
         dataset=run_ds,